Flag call to deprecated `get_theme()`
I don't think this is actually used. A search for the
`bucc_mobile_stylesheet()` function - which makes the call to
`get_theme()` - in the BU IS&T Github organization seems to confirm
that it is only called within the callback for the
`bucc_current_stylesheet` filter. A search for that filter seems to
confirm that it's not actually applied anywhere.