refactor the write file function to use S3
directly. The filesystem based fwrite() functions should work in theory with the S3 Upload plugin, but in practice they do a lot of reading and moving files around from tmp and I'm not surprised they don't work given the complexity. This new function definition just uses the AWS PHP SDK connection directly from the bu-media-s3 plugin and writes the custom css file to the bucket instead of going through the filesystem calls. This will only work with the bu-media-s3 plugin, and will only write to files in S3.