hard-code the css filenames
instead of relying on pulling the filenames from the existing files on disk.  Once the files are in S3, asking for their filenames again adds a lot of complexity and actually the files have always been called custom.css and custom.min.css. It seems easier to just assume that we can keep calling them that.