skip file if the minified css is below the inline threshold
currently 2000 characters. Stylesheets with less than 200 characters will be add directly inline with the html, so no external file is needed. This does raise a use case of stylesheets that were above 2000 but drop below. In that case there will be an unreferenced file left behind. That maybe okay, ultimately it may be better to switch to the post WP 4.7 core css customizations, which appears to always add custom css inline anyway. Maybe we don't need to be fancier than core.